test(ui): move specs to tests dir, use i18next cimode, fix Title hook order
